pyproject.toml
[tool.poetry]
name = "python-django"
version = "0.1.0"
description = ""
authors = ["Кутюрин Павел  <paulku@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
django-filter = "^23.3"
django-debug-toolbar = "^4.2.0"
pillow = "^10.0.1"
drf-spectacular = "^0.26.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
